@@ -2,8 +2,6 @@ import 'dart:io';
 import 'dart:async';
 import 'dart:convert';
 
-import 'package:aws_lambda_dart_runtime/runtime/environment.dart';
-import 'package:meta/meta.dart';
 import 'package:http/http.dart' as http;
 
 /// Next invocation data wraps the data from the
@@ -24,26 +22,26 @@ class NextInvocation {
   final String requestId;
 
   /// Deadline milliseconds is the setting for ultimate cancelation of the invocation.
-  final String deadlineMs;
+  final String? deadlineMs;
 
   /// Invoked function ARN is the identifier of the function.
-  final String invokedFunctionArn;
+  final String? invokedFunctionArn;
 
   /// Tracing id is the identifier for tracing like X-Ray.
-  final String traceId;
+  final String? traceId;
 
   /// Client context is the context that is provided to the function.
-  final String clientContext;
+  final String? clientContext;
 
   /// Cognito identity is the identity that maybe is used for authorizing the request.
-  final String cognitoIdentity;
+  final String? cognitoIdentity;
 
   /// Digesting a [HttpClientResponse] into a [NextInvocation].
   static Future<NextInvocation> fromResponse(http.Response response) async =>
       NextInvocation(
           response: (json.decode(utf8.decode(response.bodyBytes)) as Map)
               .cast<String, dynamic>(),
-          requestId: response.headers[runtimeRequestId],
+          requestId: response.headers[runtimeRequestId]!,
           deadlineMs: response.headers[runtimeDeadlineMs],
           invokedFunctionArn: response.headers[runtimeInvokedFunctionArn],
           traceId: response.headers[runtimeTraceId],
@@ -51,14 +49,14 @@ class NextInvocation {
           cognitoIdentity: response.headers[runtimeCognitoIdentity]);
 
   const NextInvocation({
-    @required this.requestId,
+    required this.requestId,
     this.deadlineMs,
     this.traceId,
     this.clientContext,
     this.cognitoIdentity,
     this.invokedFunctionArn,
-    this.response,
-  }) : assert(requestId != null);
+    required this.response,
+  });
 }
 
 /// Invocation result is the result that the invoked handler
@@ -73,9 +71,7 @@ class InvocationResult {
   /// any json-encodable data type.
   final dynamic body;
 
-  const InvocationResult(this.requestId, this.body)
-      : assert(requestId != null),
-        assert(body != null);
+  const InvocationResult(this.requestId, this.body) : assert(body != null);
 }
 
 /// Invocation error occurs when there has been an
@@ -104,32 +100,38 @@ class InvocationError {
 /// It is implemented as a singleton whereby [Client.instance]
 /// always returns the already instantiated client.
 class Client {
-  final http.Client _client = http.Client();
-  final Environment environment;
+  late http.Client _client;
 
-  factory Client({Environment? environment}) {
-    environment ??= Environment();
+  static final Client _singleton = Client._internal();
 
-    return Client.raw(environment: environment);
+  factory Client() {
+    return _singleton;
   }
 
-  Client.raw({required this.environment}) : assert(environment != null);
+  Client._internal() {
+    _client = http.Client();
+  }
 
-  static const _kruntimeApiVersion = '2018-06-01';
+  static const runtimeApiVersion = '2018-06-01';
+
+  static String? get runtimeApi =>
+      Platform.environment['AWS_LAMBDA_RUNTIME_API'];
+
+  static String? get handlerName => Platform.environment['_HANDLER'];
 
   /// Get the next invocation from the AWS Lambda Runtime Interface (see https://docs.aws.amazon.com/lambda/latest/dg/runtimes-api.html).
   Future<NextInvocation> getNextInvocation() async {
     final response = await _client.get(Uri.parse(
-        'http://${environment.runtimeAPI}//$_kruntimeApiVersion/invocation/next'));
+        'http://$runtimeApi/$runtimeApiVersion/runtime/invocation/next'));
     return NextInvocation.fromResponse(response);
   }
 
   /// Post the invocation response to the AWS Lambda Runtime Interface.
   Future<http.Response> postInvocationResponse(
-      String requestId, dynamic payload) async {
+      String? requestId, dynamic payload) async {
     return await _client.post(
       Uri.parse(
-        'http://${environment.runtimeAPI}/$_kruntimeApiVersion/runtime/invocation/$requestId/response',
+        'http://$runtimeApi/$runtimeApiVersion/runtime/invocation/$requestId/response',
       ),
       body: jsonEncode(payload),
     );
@@ -142,7 +144,7 @@ class Client {
       String requestId, InvocationError err) async {
     return await _client.post(
       Uri.parse(
-        'http://${environment.runtimeAPI}/$_kruntimeApiVersion/runtime/invocation/$requestId/error',
+        'http://$runtimeApi/$runtimeApiVersion/runtime/invocation/$requestId/error',
       ),
       body: jsonEncode(err),
       headers: {'Content-type': 'application/json'},
