@@ -12,32 +12,32 @@ class AwsALBEvent extends Event {
   /// Request context in which this request is executed.
   /// For the ELB this is the ARN of the target group.
   @JsonKey()
-  final AwsALBEventContext context;
+  final AwsALBEventContext? context;
 
   /// HTTP method that is used to trigger the invocation of the Lambda.
   @JsonKey()
-  final String httpMethod;
+  final String? httpMethod;
 
   /// The URI that is accessed to trigger the invocation of the Lambda.
   @JsonKey()
-  final String path;
+  final String? path;
 
   /// HTTP headers that are send with the request to the load balancer.
   @JsonKey()
-  final Map<String, dynamic> headers;
+  final Map<String, dynamic>? headers;
 
   /// The query parameters for the request to the load balancer.
   @JsonKey()
-  final Map<String, dynamic> queryStringParameters;
+  final Map<String, dynamic>? queryStringParameters;
 
   /// Body of the request. This can be data that is send with the POST
   /// to the request.
   @JsonKey()
-  final String body;
+  final String? body;
 
   /// Singals that the request is Base64 encoded.
   @JsonKey()
-  final bool isBase64Encoded;
+  final bool? isBase64Encoded;
 
   factory AwsALBEvent.fromJson(Map<String, dynamic> json) =>
       _$AwsALBEventFromJson(json);
@@ -60,20 +60,20 @@ class AwsALBEvent extends Event {
 /// They should reflect the informationen needed here.
 class AwsALBResponse {
   /// The body of the HTTP Response send from the API Gateway to the client.
-  String body = '';
+  String? body;
 
   /// Indicates if the [body] is Base64 encoded or not. By default is `false`.
-  bool isBase64Encoded = false;
+  bool? isBase64Encoded;
 
   /// HTTP status code of the response of the API Gateway to the client.
   /// The default status code is `200 OK`.
-  int statusCode = HttpStatus.ok;
+  int? statusCode;
 
   /// Description of the send HTTP status code.
-  String statusDescription = '200 OK';
+  String? statusDescription;
 
   /// The HTTP headers that should be send with the response to the client.
-  Map<String, String> headers = {'Content-Type': 'text/html; charset=utf-8'};
+  Map<String, String>? headers;
 
   /// Returns the JSON representation of the response. This is called by
   /// the JSON encoder to produce the response.
@@ -87,10 +87,10 @@ class AwsALBResponse {
 
   factory AwsALBResponse.fromString(
     String body, {
-    isBase64Encoded,
-    statusCode,
-    statusDescription,
-    headers,
+    bool? isBase64Encoded,
+    int? statusCode,
+    String? statusDescription,
+    Map<String, String>? headers,
   }) {
     return AwsALBResponse(
         body: body,
@@ -103,12 +103,18 @@ class AwsALBResponse {
   /// The Response that should be returned to the Application Load Balancer.
   /// It is constructed with some default values for the optional parameters.
   AwsALBResponse({
-    body,
-    headers,
-    isBase64Encoded,
-    statusCode,
-    statusDescription,
-  });
+    String? body,
+    Map<String, String>? headers,
+    bool? isBase64Encoded,
+    int? statusCode,
+    String? statusDescription,
+  }) {
+    this.body = body ?? '';
+    this.isBase64Encoded = isBase64Encoded ?? false;
+    this.headers = headers ?? {'Content-Type': 'text/html; charset=utf-8'};
+    this.statusCode = statusCode ?? HttpStatus.ok;
+    this.statusDescription = statusDescription ?? '200 OK';
+  }
 }
 
 /// AWS ALB Event Request Context ...
