@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:aws_lambda_dart_runtime/aws_lambda_dart_runtime.dart';
 
 import 'package:aws_lambda_dart_runtime/client/client.dart';
-import 'package:aws_lambda_dart_runtime/runtime/environment.dart';
 import 'package:aws_lambda_dart_runtime/runtime/event.dart';
 import 'package:aws_lambda_dart_runtime/runtime/context.dart';
 import 'package:aws_lambda_dart_runtime/runtime/exception.dart';
@@ -17,9 +16,7 @@ class _RuntimeHandler {
   final Type type;
   final dynamic handler;
 
-  const _RuntimeHandler(this.type, this.handler)
-      : assert(type != null),
-        assert(handler != null);
+  const _RuntimeHandler(this.type, this.handler) : assert(handler != null);
 }
 
 /// A Runtime manages the interface to the Lambda API.
@@ -40,26 +37,23 @@ class _RuntimeHandler {
 ///
 /// Note: You can register an
 class Runtime {
-  late final Client _client;
-  late final Environment _env;
+  late Client _client;
 
-  static final Runtime _instance = Runtime._internal();
-  factory Runtime() => _instance;
-
+  static final Runtime _singleton = Runtime._internal();
   final Map<String, _RuntimeHandler> _handlers = {};
 
+  factory Runtime() {
+    return _singleton;
+  }
+
   Runtime._internal() {
-    _env = Environment();
-    _client = Client(environment: _env);
+    _client = Client();
   }
 
   /// Lists the registered handlers by name.
   /// The name is a simple [String] which reflects
   /// the name of the trigger in the Lambda Execution API.
   List<String> get handlers => _handlers.keys.toList();
-
-  /// Get the runtime environment.
-  Environment get environment => _env;
 
   /// Checks if a specific handlers has been registered
   /// with the runtime.
@@ -74,8 +68,8 @@ class Runtime {
   }
 
   /// Unregister a handler function [Handler] with [name].
-  Handler deregisterHandler(String name) =>
-      _handlers.remove(name)?.handler as Handler;
+  Handler? deregisterHandler(String name) =>
+      _handlers.remove(name)?.handler as Handler?;
 
   /// Register an new event to be ingested by a handler.
   /// The type should reflect your type in your handler definition [Handler<T>].
@@ -95,20 +89,20 @@ class Runtime {
   /// If there is an error during the execution. The exception gets caught
   /// and the error is posted via [_client.postInvocationError(err)] to the API.
   void invoke() async {
-    do {
-      _handleInvocation(await _client.getNextInvocation());
-    } while (true);
+    while (true) {
+      await _handleInvocation(await _client.getNextInvocation());
+    }
   }
 
-  void _handleInvocation(NextInvocation nextInvocation) async {
+  Future<void> _handleInvocation(NextInvocation nextInvocation) async {
     try {
       // creating the new context
       final context = Context.fromNextInvocation(nextInvocation);
 
-      final func = _handlers[environment.handler];
+      final func = _handlers[context.handler];
       if (func == null) {
         throw RuntimeException(
-            'No handler with name "${environment.handler}" registered in runtime!');
+            'No handler with name "${context.handler}" registered in runtime!');
       }
       final event = Event.fromHandler(func.type, nextInvocation.response);
       final result = await func.handler(context, event);
