@@ -8,9 +8,9 @@ part of 'alexa_event.dart';
 
 AwsAlexaEventHeader _$AwsAlexaEventHeaderFromJson(Map<String, dynamic> json) {
   return AwsAlexaEventHeader(
-    namespace: json['namespace'] as String,
-    payloadVersion: json['payloadVersion'] as String,
-    name: json['name'] as String,
+    namespace: json['namespace'] as String?,
+    payloadVersion: json['payloadVersion'] as String?,
+    name: json['name'] as String?,
   );
 }
 
@@ -24,9 +24,10 @@ Map<String, dynamic> _$AwsAlexaEventHeaderToJson(
 
 AwsAlexaEvent _$AwsAlexaEventFromJson(Map<String, dynamic> json) {
   return AwsAlexaEvent(
-    header:
-        AwsAlexaEventHeader.fromJson(json['header'] as Map<String, dynamic>),
-    payload: json['payload'] as Map<String, dynamic>,
+    header: json['header'] == null
+        ? null
+        : AwsAlexaEventHeader.fromJson(json['header'] as Map<String, dynamic>),
+    payload: json['payload'] as Map<String, dynamic>?,
   );
 }
 
