@@ -9,14 +9,14 @@ part of 'cloudwatch_event.dart';
 AwsCloudwatchEvent _$AwsCloudwatchEventFromJson(Map<String, dynamic> json) {
   return AwsCloudwatchEvent(
     resources:
-        (json['resources'] as List<dynamic>).map((e) => e as String).toList(),
-    region: json['region'] as String,
-    id: json['id'] as String,
-    source: json['source'] as String,
-    account: json['account'] as String,
-    detailType: json['detail-type'] as String,
-    detail: json['detail'] as Map<String, dynamic>,
-    time: DateTime.parse(json['time'] as String),
+        (json['resources'] as List<dynamic>?)?.map((e) => e as String).toList(),
+    region: json['region'] as String?,
+    id: json['id'] as String?,
+    source: json['source'] as String?,
+    account: json['account'] as String?,
+    detailType: json['detail-type'] as String?,
+    detail: json['detail'] as Map<String, dynamic>?,
+    time: json['time'] == null ? null : DateTime.parse(json['time'] as String),
   );
 }
 
@@ -29,5 +29,5 @@ Map<String, dynamic> _$AwsCloudwatchEventToJson(AwsCloudwatchEvent instance) =>
       'account': instance.account,
       'detail-type': instance.detailType,
       'detail': instance.detail,
-      'time': instance.time.toIso8601String(),
+      'time': instance.time?.toIso8601String(),
     };
