@@ -9,7 +9,7 @@ part of 'cloudwatch_log_event.dart';
 AwsCloudwatchLogEvent _$AwsCloudwatchLogEventFromJson(
     Map<String, dynamic> json) {
   return AwsCloudwatchLogEvent(
-    awslogs: json['awslogs'] as Map<String, dynamic>,
+    awslogs: json['awslogs'] as Map<String, dynamic>?,
   );
 }
 
