@@ -9,9 +9,9 @@ part of 'dynamodb_event.dart';
 AwsDynamoDBUpdateRecord _$AwsDynamoDBUpdateRecordFromJson(
     Map<String, dynamic> json) {
   return AwsDynamoDBUpdateRecord(
-    keys: json['Keys'] as Map<String, dynamic>,
-    oldImage: json['NewImage'] as Map<String, dynamic>,
-    newImage: json['OldImage'] as Map<String, dynamic>,
+    keys: json['Keys'] as Map<String, dynamic>?,
+    oldImage: json['NewImage'] as Map<String, dynamic>?,
+    newImage: json['OldImage'] as Map<String, dynamic>?,
   );
 }
 
@@ -26,12 +26,12 @@ Map<String, dynamic> _$AwsDynamoDBUpdateRecordToJson(
 AwsDynamoDBUpdateEventRecord _$AwsDynamoDBUpdateEventRecordFromJson(
     Map<String, dynamic> json) {
   return AwsDynamoDBUpdateEventRecord(
-    eventId: json['eventId'] as String,
-    eventName: json['eventName'] as String,
-    eventSource: json['eventSource'] as String,
-    eventVersion: json['eventVersion'] as String,
-    awsRegion: json['awsRegion'] as String,
-    eventSourceARN: json['eventSourceARN'] as String,
+    eventId: json['eventId'] as String?,
+    eventName: json['eventName'] as String?,
+    eventSource: json['eventSource'] as String?,
+    eventVersion: json['eventVersion'] as String?,
+    awsRegion: json['awsRegion'] as String?,
+    eventSourceARN: json['eventSourceARN'] as String?,
   );
 }
 
@@ -49,8 +49,8 @@ Map<String, dynamic> _$AwsDynamoDBUpdateEventRecordToJson(
 AwsDynamoDBUpdateEvent _$AwsDynamoDBUpdateEventFromJson(
     Map<String, dynamic> json) {
   return AwsDynamoDBUpdateEvent(
-    records: (json['Records'] as List<dynamic>)
-        .map((e) =>
+    records: (json['Records'] as List<dynamic>?)
+        ?.map((e) =>
             AwsDynamoDBUpdateEventRecord.fromJson(e as Map<String, dynamic>))
         .toList(),
   );
