@@ -8,8 +8,8 @@ part of 's3_event.dart';
 
 AwsS3Event _$AwsS3EventFromJson(Map<String, dynamic> json) {
   return AwsS3Event(
-    records: (json['Records'] as List<dynamic>)
-        .map((e) => AwsS3EventRecord.fromJson(e as Map<String, dynamic>))
+    records: (json['Records'] as List<dynamic>?)
+        ?.map((e) => AwsS3EventRecord.fromJson(e as Map<String, dynamic>))
         .toList(),
   );
 }
@@ -21,17 +21,27 @@ Map<String, dynamic> _$AwsS3EventToJson(AwsS3Event instance) =>
 
 AwsS3EventRecord _$AwsS3EventRecordFromJson(Map<String, dynamic> json) {
   return AwsS3EventRecord(
-    eventName: json['eventName'] as String,
-    eventSource: json['eventSource'] as String,
-    awsRegion: json['awsRegion'] as String,
-    eventTime: DateTime.parse(json['eventTime'] as String),
-    eventVersion: json['eventVersion'] as String,
-    userIdentity: AwsS3UserIdentity.fromJson(
-        json['userIdentity'] as Map<String, dynamic>),
+    eventName: json['eventName'] as String?,
+    eventSource: json['eventSource'] as String?,
+    awsRegion: json['awsRegion'] as String?,
+    eventTime: json['eventTime'] == null
+        ? null
+        : DateTime.parse(json['eventTime'] as String),
+    eventVersion: json['eventVersion'] as String?,
+    userIdentity: json['userIdentity'] == null
+        ? null
+        : AwsS3UserIdentity.fromJson(
+            json['userIdentity'] as Map<String, dynamic>),
     requestParameters:
-        Map<String, String>.from(json['requestParameters'] as Map),
-    responseElements: Map<String, String>.from(json['responseElements'] as Map),
-    s3: AwsS3Data.fromJson(json['s3'] as Map<String, dynamic>),
+        (json['requestParameters'] as Map<String, dynamic>?)?.map(
+      (k, e) => MapEntry(k, e as String),
+    ),
+    responseElements: (json['responseElements'] as Map<String, dynamic>?)?.map(
+      (k, e) => MapEntry(k, e as String),
+    ),
+    s3: json['s3'] == null
+        ? null
+        : AwsS3Data.fromJson(json['s3'] as Map<String, dynamic>),
   );
 }
 
@@ -40,7 +50,7 @@ Map<String, dynamic> _$AwsS3EventRecordToJson(AwsS3EventRecord instance) =>
       'eventVersion': instance.eventVersion,
       'eventSource': instance.eventSource,
       'awsRegion': instance.awsRegion,
-      'eventTime': instance.eventTime.toIso8601String(),
+      'eventTime': instance.eventTime?.toIso8601String(),
       'eventName': instance.eventName,
       'userIdentity': instance.userIdentity,
       'requestParameters': instance.requestParameters,
@@ -50,10 +60,14 @@ Map<String, dynamic> _$AwsS3EventRecordToJson(AwsS3EventRecord instance) =>
 
 AwsS3Data _$AwsS3DataFromJson(Map<String, dynamic> json) {
   return AwsS3Data(
-    s3SchemaVersion: json['s3SchemaVersion'] as String,
-    configurationId: json['configurationId'] as String,
-    bucket: AwsS3Bucket.fromJson(json['bucket'] as Map<String, dynamic>),
-    object: AWSS3EventObject.fromJson(json['object'] as Map<String, dynamic>),
+    s3SchemaVersion: json['s3SchemaVersion'] as String?,
+    configurationId: json['configurationId'] as String?,
+    bucket: json['bucket'] == null
+        ? null
+        : AwsS3Bucket.fromJson(json['bucket'] as Map<String, dynamic>),
+    object: json['object'] == null
+        ? null
+        : AWSS3EventObject.fromJson(json['object'] as Map<String, dynamic>),
   );
 }
 
@@ -66,10 +80,12 @@ Map<String, dynamic> _$AwsS3DataToJson(AwsS3Data instance) => <String, dynamic>{
 
 AwsS3Bucket _$AwsS3BucketFromJson(Map<String, dynamic> json) {
   return AwsS3Bucket(
-    name: json['name'] as String,
-    ownerIdentity: AwsS3UserIdentity.fromJson(
-        json['ownerIdentity'] as Map<String, dynamic>),
-    arn: json['arn'] as String,
+    name: json['name'] as String?,
+    ownerIdentity: json['ownerIdentity'] == null
+        ? null
+        : AwsS3UserIdentity.fromJson(
+            json['ownerIdentity'] as Map<String, dynamic>),
+    arn: json['arn'] as String?,
   );
 }
 
@@ -82,10 +98,10 @@ Map<String, dynamic> _$AwsS3BucketToJson(AwsS3Bucket instance) =>
 
 AWSS3EventObject _$AWSS3EventObjectFromJson(Map<String, dynamic> json) {
   return AWSS3EventObject(
-    key: json['key'] as String,
-    size: json['size'] as int,
-    eTag: json['eTag'] as String,
-    sequencer: json['sequencer'] as String,
+    key: json['key'] as String?,
+    size: json['size'] as int?,
+    eTag: json['eTag'] as String?,
+    sequencer: json['sequencer'] as String?,
   );
 }
 
@@ -99,7 +115,7 @@ Map<String, dynamic> _$AWSS3EventObjectToJson(AWSS3EventObject instance) =>
 
 AwsS3UserIdentity _$AwsS3UserIdentityFromJson(Map<String, dynamic> json) {
   return AwsS3UserIdentity(
-    principalId: json['principalId'] as String,
+    principalId: json['principalId'] as String?,
   );
 }
 
