@@ -8,15 +8,15 @@ part of 'sqs_event.dart';
 
 AwsSQSEventRecord _$AwsSQSEventRecordFromJson(Map<String, dynamic> json) {
   return AwsSQSEventRecord(
-    messageId: json['messageId'] as String,
-    body: json['body'] as String,
-    receiptHandle: json['receiptHandle'] as String,
-    attributes: json['attributes'] as Map<String, dynamic>,
-    messageAttributes: json['messageAttributes'] as Map<String, dynamic>,
-    md5OfBody: json['md5OfBody'] as String,
-    eventSource: json['eventSource'] as String,
-    eventSourceARN: json['eventSourceARN'] as String,
-    awsRegion: json['awsRegion'] as String,
+    messageId: json['messageId'] as String?,
+    body: json['body'] as String?,
+    receiptHandle: json['receiptHandle'] as String?,
+    attributes: json['attributes'] as Map<String, dynamic>?,
+    messageAttributes: json['messageAttributes'] as Map<String, dynamic>?,
+    md5OfBody: json['md5OfBody'] as String?,
+    eventSource: json['eventSource'] as String?,
+    eventSourceARN: json['eventSourceARN'] as String?,
+    awsRegion: json['awsRegion'] as String?,
   );
 }
 
@@ -35,8 +35,8 @@ Map<String, dynamic> _$AwsSQSEventRecordToJson(AwsSQSEventRecord instance) =>
 
 AwsSQSEvent _$AwsSQSEventFromJson(Map<String, dynamic> json) {
   return AwsSQSEvent(
-    records: (json['Records'] as List<dynamic>)
-        .map((e) => AwsSQSEventRecord.fromJson(e as Map<String, dynamic>))
+    records: (json['Records'] as List<dynamic>?)
+        ?.map((e) => AwsSQSEventRecord.fromJson(e as Map<String, dynamic>))
         .toList(),
   );
 }
