@@ -8,24 +8,25 @@ import 'package:test/test.dart';
 final file = 'data/apigateway_event.json';
 
 final String contents = File(file).readAsStringSync();
-final Map<String, dynamic> json = jsonDecode(contents) as Map<String, dynamic>;
+final Map<String, dynamic>? json =
+    jsonDecode(contents) as Map<String, dynamic>?;
 
 void main() {
   group('apigateway_default', () {
     test('json got parsed and creates an event', () async {
-      final event = AwsApiGatewayEvent.fromJson(json);
+      final event = AwsApiGatewayEvent.fromJson(json!);
 
       expect(event.body, equals(jsonEncode({'foo': 'bar'})));
       expect(event.path, equals('/test/hello'));
-      expect(event.headers.acceptEncoding,
+      expect(event.headers!.acceptEncoding,
           equals('gzip, deflate, lzma, sdch, br'));
-      expect(event.requestContext.httpMethod, equals('POST'));
-      expect(event.requestContext.accountId, equals('123456789012'));
-      expect(event.requestContext.requestId,
+      expect(event.requestContext!.httpMethod, equals('POST'));
+      expect(event.requestContext!.accountId, equals('123456789012'));
+      expect(event.requestContext!.requestId,
           equals('41b45ea3-70b5-11e6-b7bd-69b5aaebc7d9'));
       expect(event.queryStringParameters, equals({'name': 'me'}));
-      expect(event.requestContext.resourcePath, equals('/{proxy+}'));
-      expect(event.headers.raw['Accept-Encoding'],
+      expect(event.requestContext!.resourcePath, equals('/{proxy+}'));
+      expect(event.headers!.raw['Accept-Encoding'],
           equals('gzip, deflate, lzma, sdch, br'));
     });
     test('factory creates event with default values', () {
