@@ -14,14 +14,14 @@ void main() {
   group('runtime', () {
     test('Custom event is add to the events', () {
       Event.registerEvent<CustomTestEvent>(
-          (Map<String, dynamic> json) => CustomTestEvent.fromJson({}));
+          (Map<String, dynamic>? json) => CustomTestEvent.fromJson({}));
 
       expect(Event.exists<CustomTestEvent>(), true);
     });
 
     test('Custom event is deregistered', () {
       Event.registerEvent<CustomTestEvent>(
-          (Map<String, dynamic> json) => CustomTestEvent.fromJson({}));
+          (Map<String, dynamic>? json) => CustomTestEvent.fromJson({}));
 
       Event.deregisterEvent<CustomTestEvent>();
 
